feat(evm): account state sync (#14)
* return account updates on commit

* write account updates to wallet store

* update abi

* events module

* parse Voted/Unvoted events

* update wallet attributes

* Cargo.lock

* cleanup

* fix cast

* use BTreeMap to preserve insertion order

+++ packages/api-database/source/repositories/wallet-repository.ts
@@ -19,7 +19,7 @@ export const makeWalletRepository = (dataSource: RepositoryDataSource): WalletRe
 				convertToJsonbSorting(sorting, [
 					{
 						direction: "desc",
-						jsonFieldAccessor: { cast: "bigint", fieldName: "balance", operator: "->>" },
+						jsonFieldAccessor: { cast: "numeric", fieldName: "balance", operator: "->>" },
 						property: "attributes",
 					},
 				]),
